LayoutTests/http/tests/websocket/tests/hybi/interleaved-fragments-expected.txt

CONSOLE MESSAGE: Received new data frame but previous continuous frame is unfinished.
Test whether WebSocket rejects interleaved fragmented frames.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

onopen() was called.
PASS closeEvent.wasClean is false
PASS successfullyParsed is true

TEST COMPLETE
